Clean refactored project structure
