--- backend/MonChenil.Tests/GlobalUsings.cs
@@ -1,1 +1,0 @@
-global using Xunit;
